simple calendar at front page
