+++ date/views.py
@@ -10,13 +10,21 @@ from news.models import Post
 
 from itertools import chain
 
+from django.views import generic
+from django.utils.safestring import mark_safe
+from events.models import *
+from events.utils import Calendar
 
 def index(request):
     events = Event.objects.filter(published=True, event_date_end__gte=datetime.date.today()).order_by('event_date_start')
     news = Post.objects.filter(published=True).reverse()[:2]
     news_events = list(chain(events, news))
 
-    return render(request, 'date/start.html', {'news_events': news_events, 'events': events, 'news': news})
+    d = get_date(request.GET.get('day', None))
+    cal = Calendar(d.year, d.month)
+    calendar = cal.formatmonth(withyear=True)
+
+    return render(request, 'date/start.html', {'news_events': news_events, 'events': events, 'news': news, 'calendar': calendar })
 
 
 def language(request, lang):
@@ -42,3 +50,9 @@ def handler500(request, *args, **argv):
     response = render(request, '500.html', {})
     response.status_code = 404
     return response
+
+def get_date(req_day):
+    if req_day:
+        year, month = (int(x) for x in req_day.split('-'))
+        return date(year, month, day=1)
+    return datetime.date.today()
