+++ events/views.py
@@ -7,11 +7,7 @@ from django.shortcuts import render
 from .models import Event, EventAttendees
 from websocket import create_connection
 from websocket._exceptions import WebSocketBadStatusException
-from django.views import generic
-from django.utils.safestring import mark_safe
 
-from .models import *
-from .utils import Calendar
 import json
 
 import logging
@@ -87,25 +83,3 @@ def ws_data(form):
     del pref['anonymous']
     del pref['email']
     return {"data": pref}
-
-class CalendarView(generic.ListView):
-    model = Event
-    template_name = 'events/calendar.html'
-
-    def get_context_data(self, **kwargs):
-        context = super(CalendarView, self).get_context_data(**kwargs)
-
-        d = get_date(self.request.GET.get('day', None))
-        
-        cal = Calendar(d.year, d.month)
-
-        html_cal = cal.formatmonth(withyear=True)
-        context['calendar'] = mark_safe(html_cal)
-
-        return context
-
-def get_date(req_day):
-    if req_day:
-        year, month = (int(x) for x in req_day.split('-'))
-        return date(year, month, day=1)
-    return datetime.date.today()
